Refactor ServiceLocator method names and improve Vulkan texture handling
- Changed all occurrences of `getService` to `GetService` in the ServiceLocator class and its usage throughout the codebase for consistency.
- Added a temporary texture member in VulkanRenderer for future texture management.
- Updated Vulkan texture handling to support multiple image layers in Texture class.
- Enhanced TextureAtlas to load multiple textures and manage their regions effectively.
- Adjusted Vulkan command buffer and image view creation to accommodate new texture features.
- Fixed various minor issues and improved code readability across Vulkan-related files.

+++ src/vulkan/data/data.h
@@ -34,7 +34,7 @@ struct InstanceDataTwoD {
   glm::vec2 pos;
   glm::vec2 dim;
   float rot;
-  int32_t textureIndex;
+  float textureIndex;
   glm::vec4 atlasCoords;
 };
